requirements.txt
python-telegram-bot==20.7
google-generativeai>=0.7.2
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
streamlit==1.30.0
